support state sets
Renders labeled gauges with 0 or 1 for the different possible states.
Build on top of other float64 code, which is a little unfortunate
because now I need to do float64 math. Probably want the state gauges to
follow a different path with `[]byte`s.

refs #11

+++ internal/config/config.go
@@ -41,8 +41,8 @@ func (m *Modbus) Defaults() {
 }
 
 type State struct {
-	Name  string `json:"name,omitempty" yaml:"name,omitempty"`
-	Value uint16 `json:"value,omitempty" yaml:"value,omitempty"`
+	Name  string  `json:"name,omitempty" yaml:"name,omitempty"`
+	Value float64 `json:"value,omitempty" yaml:"value,omitempty"`
 }
 
 type Gauge struct {
